LayoutTests/fast/zooming/client-rect-in-fixed-zoomed-expected.txt

Tests localToAbsolute coordinate conversion for an element inside position:fixed after zoom.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.pageScaleFactor() is 2
PASS JSON.stringify(clientRect) is JSON.stringify({top: -108, right: 42, bottom: -98, left: 22, width: 20, height: 10})
PASS successfullyParsed is true

TEST COMPLETE
